Struct pica::commands::filter::Filter

pub(crate) struct Filter {

    skip_invalid: bool,
    invert_match: bool,
    ignore_case: bool,
    strsim_threshold: u8,
    keep: Option<String>,
    discard: Option<String>,
    expr_file: Option<PathBuf>,
    allow_lists: Vec<PathBuf>,
    deny_lists: Vec<PathBuf>,
    limit: usize,
    and: Vec<String>,
    or: Vec<String>,
    not: Vec<String>,
    gzip: bool,
    append: bool,
    tee: Option<PathBuf>,
    progress: bool,
    output: Option<OsString>,
    filter: String,
    filenames: Vec<OsString>,

}

Filter records by whether the given filter expression matches

Fields
skip_invalid: bool

Skip invalid records that can’t be decoded as normalized PICA+

invert_match: bool

Filter only records that did not match

ignore_case: bool

When this flag is provided, comparison operations will be search case insensitive

strsim_threshold: u8

The minimum score for string similarity comparisons (range: 0.0..1.0)

keep: Option<String>

Keep only fields specified by a list of predicates.

discard: Option<String>

Discard fields specified by a list of predicates.

expr_file: Option<PathBuf>

Take a filter expression from <EXPR_FILE>

Note: Do not provide an additional filter expression as an CLI argument!

allow_lists: Vec<PathBuf>

Ignore records which are not explicitly listed in one of the given allow-lists.

A allow-list must be an CSV/TSV or Apache Arrow file, whereby a column idn exists. If the file extension is .feather, .arrow, or .ipc the file is automatically interpreted as Apache Arrow; file existions .csv, .csv.gz, .tsv or .tsv.gz is interpreted as CSV/TSV.

deny_lists: Vec<PathBuf>

Ignore records which are explicitly listed in one of the given deny-lists.

A deny-list must be an CSV/TSV or Apache Arrow file, whereby a column idn exists. If the file extension is .feather, .arrow, or .ipc the file is automatically interpreted as Apache Arrow; file existions .csv, .csv.gz, .tsv or .tsv.gz is interpreted as CSV/TSV.

limit: usize

Limit the result to first records

Note: A limit value 0 means no limit.

and: Vec<String>

Connects the filter with additional expressions using the logical AND-operator (conjunction)

This option can’t be combined with --or or --not.

or: Vec<String>

Connects the filter with additional expressions using the logical OR-operator (disjunction)

This option can’t be combined with --and or --not.

not: Vec<String>

Connects the filter with additional expressions using the logical NOT-operator (negation)

This option can’t be combined with --and or --or.

gzip: bool

Compress output in gzip format

append: bool

Append to the given file, do not overwrite

Warning: This option can’t be used when writing to a gzip file.

tee: Option<PathBuf>

Write simultaneously to the file and stdout

progress: bool

Show progress bar (requires -o/--output).

output: Option<OsString>

Write output to instead of stdout

filter: String

A filter expression used for searching

filenames: Vec<OsString>

Read one or more files in normalized PICA+ format

If no filenames where given or a filename is “-”, data is read from standard input (stdin).

Implementations

impl Filter

pub(crate) fn run(self, config: &Config) -> Result<(), CliError>