Struct pica::ReaderBuilder

pub struct ReaderBuilder { /* fields omitted */ }

Configures and builds a PICA+ reader.

Implementations

impl ReaderBuilder

pub fn new() -> ReaderBuilder

Create a new ReaderBuilder for reading PICA+ data.

Example

use pica::{ByteRecord, ReaderBuilder, StringRecord};
use std::error::Error;

fn example() -> Result<(), Box<dyn Error>> {
    let data = "003@ \x1f0123456789\x1e\n";
    let mut reader = ReaderBuilder::new().from_reader(data.as_bytes());
    let records = reader
        .records()
        .map(Result::unwrap)
        .collect::<Vec<StringRecord>>();

    assert_eq!(
        records,
        vec![StringRecord::from_byte_record(ByteRecord::from_bytes(
            "003@ \x1f0123456789\x1e\n".as_bytes()
        )?)?]
    );

    Ok(())
}

pub fn from_path<P: AsRef<Path>>(
&self,
path: P
) -> Result<Reader<Box<dyn Read>>>

Builds a new Reader with the current configuration, that reads from a file path.

Example

use pica::ReaderBuilder;
use std::error::Error;

fn example() -> Result<(), Box<dyn Error>> {
    let mut reader =
        ReaderBuilder::new().from_path("tests/data/119232022.dat")?;
    let record = reader.records().next().unwrap();
    assert!(record.is_ok());

    let mut reader =
        ReaderBuilder::new().from_path("tests/data/119232022.dat.gz")?;
    let record = reader.records().next().unwrap();
    assert!(record.is_ok());

    Ok(())
}

pub fn from_reader<R: Read>(&self, reader: R) -> Reader<R>

Builds a new Reader with the current configuration, that reads from an existing reader.

Example

use pica::ReaderBuilder;
use std::error::Error;
use std::io::Cursor;

fn example() -> Result<(), Box<dyn Error>> {
    let cursor = Cursor::new(b"003@ \x1f0123456789\x1e");
    let mut reader = ReaderBuilder::new().from_reader(cursor);
    let record = reader.records().next().unwrap();
    assert!(record.is_ok());

    Ok(())
}

pub fn from_path_or_stdin<P: AsRef<Path>>(
&self,
path: Option<P>
) -> Result<Reader<Box<dyn Read>>>

Builds a new Reader with the current configuration, that reads from the given path, if some was provided, otherwise from stdin.

Example

use pica::{ByteRecord, ReaderBuilder};
use std::error::Error;
use std::path::Path;

fn example() -> Result<(), Box<dyn Error>> {
    let mut reader = ReaderBuilder::new()
        .from_path_or_stdin(Some("tests/data/119232022.dat.gz"))?;
    let record = reader.records().next().unwrap();
    assert!(record.is_ok());

    let filename: Option<String> = None;
    let mut reader = ReaderBuilder::new().from_path_or_stdin(filename)?;
    assert!(reader.records().next().is_none());

    Ok(())
}

pub fn skip_invalid(self, yes: bool) -> Self

Whether to skip invalid records or not.

By default, if an invalid record occurs the reader moves forward until the next valid record was found or the end of file is reached. When this flag is disabled (yes is set to false), the iterator item wil be an pica::Error.

Example

use pica::ReaderBuilder;
use std::error::Error;
use std::io::Cursor;

fn example() -> Result<(), Box<dyn Error>> {
    let data = b"003@ \x1f0123\x1e\n003@ \x1f!456\x1e\n003@ \x1f0789\x1e\n";

    let mut reader = ReaderBuilder::new()
        .skip_invalid(false)
        .from_reader(Cursor::new(data));
    let mut iter = reader.records();

    assert!(iter.next().unwrap().is_ok());
    assert!(iter.next().unwrap().is_err());
    assert!(iter.next().unwrap().is_ok());
    assert!(iter.next().is_none());

    let mut reader = ReaderBuilder::new()
        .skip_invalid(true)
        .from_reader(Cursor::new(data));
    let mut iter = reader.records();

    assert!(iter.next().unwrap().is_ok());
    assert!(iter.next().unwrap().is_ok());
    assert!(iter.next().is_none());

    Ok(())
}

pub fn buffer_size(self, buffer_size: usize) -> Self

Change the inital capacity of an new ByteRecord.

By default the inital capacity is set to 1024 bytes.

Example

use pica::ReaderBuilder;
use std::error::Error;
use std::io::Cursor;

fn example() -> Result<(), Box<dyn Error>> {
    let cursor = Cursor::new(b"003@ \x1f0123456789\x1e");
    let mut reader =
        ReaderBuilder::new().buffer_size(2048).from_reader(cursor);
    let record = reader.records().next().unwrap();
    assert!(record.is_ok());

    Ok(())
}

pub fn limit(self, buffer_size: usize) -> Self

Change the limit of records to read.

Example

use pica::ReaderBuilder;
use std::error::Error;
use std::io::Cursor;

fn example() -> Result<(), Box<dyn Error>> {
    let cursor =
        Cursor::new(b"003@ \x1f0123456789\x1e\n003@ \x1f0123456789\x1e\n");

    let mut reader = ReaderBuilder::new().limit(1).from_reader(cursor);
    let mut records = reader.records();

    let record = records.next().unwrap();
    assert!(record.is_ok());

    assert!(records.next().is_none());

    Ok(())
}